Backend/requirements.txt
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.26.1
RapidFuzz==3.12.1
networkx==3.4.2
nltk==3.9.1
numpy==2.2.2
scipy==1.17.0
psycopg2-binary==2.9.10
PyYAML==6.0.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
joblib==1.4.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.22.3
sqlparse==0.5.3
tqdm==4.67.1
uritemplate==4.1.1
urllib3==2.3.0
